requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Deprecated==1.2.14
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.8
langchain-ollama==0.2.0
langsmith==0.1.130
ollama==0.3.3
orjson==3.10.7
packaging==24.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==8.5.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
